Rewrite the backend to match the new app data layer

+++ src/main/java/com/example/hundredplaces/data/city/CityRepository.java
@@ -2,6 +2,6 @@ package com.example.hundredplaces.data.city;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Long> {
-    City findById(long id);
+public interface CityRepository extends JpaRepository<City, Integer> {
+    City findById(int id);
 }
