+++ src/main/java/com/example/hundredplaces/data/image/ImageRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ImageRepository extends JpaRepository<Image, Long> {
-    Image findById(long id);
-    List<Image> findAllByPlaceId(Long placeId);
+public interface ImageRepository extends JpaRepository<Image, Integer> {
+    Image findById(int id);
+    List<Image> findAllByPlaceId(int placeId);
 }
