+++ src/main/java/com/example/hundredplaces/data/place/PlaceRepository.java
@@ -2,6 +2,6 @@ package com.example.hundredplaces.data.place;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaceRepository extends JpaRepository<Place, Long> {
-    Place findById(long id);
+public interface PlaceRepository extends JpaRepository<Place, Integer> {
+    Place findById(int id);
 }
